% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataobs_coda}
\alias{dataobs_coda}
\title{Simulated observation of compositional data}
\format{A data frame with 200 rows and 4 columns
\describe{
  \item{Var1}{simulated observed variable, compositional}
  \item{Var2}{simulated observed variable, compositional}
  \item{Var3}{simulated observed variable, compositional}
  \item{Group}{Factor with levels 'Group 1' and 'Group 2'}
}}
\usage{
dataobs_coda
}
\description{
A data set of 200 simulated observations with three observed variables and two groups, compositional
}
\keyword{datasets}
