% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{force_posdef}
\alias{force_posdef}
\title{Force positive definiteness}
\usage{
force_posdef(x, verbose = T)
}
\arguments{
\item{x}{matrix}

\item{verbose}{logical, default TRUE. Should the function print the corrected eigenvalues?}
}
\value{
positive definite matrix
}
\description{
Function to force positive definiteness on a matrix.
}
\author{
Solveig Pospiech
}
