% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{generalized_mean}
\alias{generalized_mean}
\alias{generalized_mean.default}
\alias{generalized_mean.rmult}
\title{Generalized mean}
\usage{
generalized_mean(x, ...)

\method{generalized_mean}{default}(
  x,
  var,
  individual_var = matrix(0, nrow = nrow(x), ncol = ncol(x)),
  ...
)

\method{generalized_mean}{rmult}(
  x,
  var,
  individual_var = matrix(0, nrow = nrow(x), ncol = ncol(x)^2),
  ...
)
}
\arguments{
\item{x}{a matrix containing the data for which the mean should be calculated}

\item{...}{not implemented}

\item{var}{a matrix containing the corrected (estimated true) group variances}

\item{individual_var}{a matrix containing individual variances. Default is a 0 - matrix with the dimensions of x, can be used for implementing the individual uncertainties}
}
\value{
vector of lenght of ncol(x) of generalized means
}
\description{
Calculates the generalized mean of a data set by using a given group variance and individual, observation-wise variances for each observation of the data set
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for class matrix or data.frame

\item \code{rmult}: for class rmult of package 'compositions'
}}

\author{
Solveig Pospiech, K. Gerald v.d. Boogaart
}
