% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uncertainties}
\alias{uncertainties}
\title{Simulated observation uncertainties}
\format{
A data frame with 200 rows and 3 columns
\describe{
  \item{Var1}{simulated observed variable}
  \item{Var2}{simulated observed variable}
  \item{Group}{Factor with levels 'Group 1' and 'Group 2'}
}
}
\usage{
uncertainties
}
\description{
A data set of 200 simulated uncertainties with two variables and two groups, non-compositional
}
\keyword{datasets}
